refactor(components): improve image for pokemon type by adding it to component
