+++ README.md
@@ -37,7 +37,7 @@ Create a web application witch [ReactJS](https://reactjs.org/) that allows users
 - [x] Pagination
 - [x] Search for Pokémon by name or number
 - [x] Pokemon card with image, name and types
-- [ ] Modal with Pokémon details
+- [x] Modal with Pokémon details
 - [ ] Page with type of species detailed
 
 ## 🖊️ Commit Messages
